pyproject.toml
[project]
name = "mcp-server-template"
version = "0.1.0"
description = "Python MCP Server template"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Alpic" }]
keywords = ["mcp", "llm", "alpic", "python"]
dependencies = [
    "black>=25.1.0",
    "mcp",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
